refactor(spf): make relocation-pipelines a DOM-free primitive
Completes the loader-pipeline extraction: relocation-pipelines moves from
behaviors/dom to primitives/ root, fully DOM-free, so it's a genuine primitive
rather than a stray non-behavior parked under behaviors/.

Two decouplings make that honest:
- Cue: the text steps are now generic over `C extends Cue` instead of naming the
  DOM `VTTCue`. Relocation only mutates startTime/endTime, which the structural
  media/types `Cue` already covers — no new type needed.
- metadata resolver: resolveVttSegmentMetadata + TextSegmentMetadata move to a
  DOM-free media/text/resolve-vtt-metadata.ts (fetch + parseVttTimestampMap, no
  document). The document-based resolveVttSegment stays in media/dom/text.

primitives/dom is never needed. relocation's test moves to primitives/tests
(now a Node test, not Chromium) with its behavior-derive dependency replaced by
an inline no-op (only the discover steps are exercised).
